+++ sda/cmd/api/api.go
@@ -26,6 +26,7 @@ import (
 	"github.com/neicnordic/sensitive-data-archive/internal/storage/v2"
 	"github.com/neicnordic/sensitive-data-archive/internal/storage/v2/locationbroker"
 	"github.com/neicnordic/sensitive-data-archive/internal/userauth"
+	"github.com/neicnordic/sensitive-data-archive/internal/v2/inboxproject"
 	"google.golang.org/grpc"
 )
 
@@ -42,6 +43,7 @@ type API struct {
 	server      *http.Server
 	inboxReader storage.Reader
 	inboxWriter storage.Writer
+	inbox       inboxproject.Config
 	db          database.Database
 	mq          broker.Broker
 }
@@ -59,6 +61,11 @@ func run() error {
 
 	if err := config.Load(); err != nil {
 		return fmt.Errorf("failed to load config: %v", err)
+	}
+
+	inbox, err := inboxproject.Load()
+	if err != nil {
+		return fmt.Errorf("failed to load inbox project config: %v", err)
 	}
 
 	db, err := postgres.NewPostgresSQLDatabase()
@@ -136,7 +143,7 @@ func run() error {
 	}
 	defer conn.Close()
 
-	app := API{grpcClient: conn, auth: auth, enforcer: e, inboxReader: inboxReader, inboxWriter: inboxWriter, db: db, mq: mq}
+	app := API{grpcClient: conn, auth: auth, enforcer: e, inboxReader: inboxReader, inboxWriter: inboxWriter, inbox: inbox, db: db, mq: mq}
 
 	serverErr := make(chan error, 1)
 	addr := apiconfig.APIAddr()